tc-250 refactor
