+++ frontend/mulighetsrommet-veileder-flate/src/components/modal/delemodal/DelemodalActions.tsx
